backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==1.10.1
motor==3.2.0
pymongo==4.5.0
beanie==1.21.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.24.1
pymupdf==1.23.4
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-generativeai==0.3.1  
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1
bcrypt==4.0.1
email-validator==2.0.0
jinja2==3.1.2
pillow==10.1.0
python-magic==0.4.27